refactor: FixedNumberGenerator를 테스트 코드에서 정의하도록 수정
