--- src/main/java/racingCar/util/FixedNumberGenerator.java
@@ -1,15 +1,0 @@
-package racingCar.util;
-
-public class FixedNumberGenerator implements NumberGenerator {
-
-    private final int fixedNumber;
-
-    public FixedNumberGenerator(int fixedNumber) {
-        this.fixedNumber = fixedNumber;
-    }
-
-    @Override
-    public int generateNumber() {
-        return fixedNumber;
-    }
-}
